Add methods to WebSite class and some functions
Methods were added to the WebSite class that will allow a
user to get a site's header info. We are using a site's date
header as a refrence for synchronizing the time on the machine.
Functions were created to reformat a string so that it can be
used to retrieve a GMT datetime object and convert that object
into a local datetime object.

Methods added to WebSite class:
	- get_site_header: Returns a web site's headers.
	- get_header_value: Returns the site's header value.

Functions added:
	- get_sites_date: Using the get_site_header in the WebSite
	class it gets the value for a site's date header as a string.
	Then, it removes the comma in the string so that it can be
	used by datetime.strptime() to return a datetime object.
	- convert_gmt_to_local: The date value used in all website's
	date header is in GMT. This function converts a datetime
	object to be in local time zone.