Fix convert GMT function to work on Python32
The function, convert_gmt_to_local, doesn't work properly on the
lab machines which are on Python 3.2. The problem was caused because
datetime.astimezone() requires an instance of a tzinfo subclass. This
was ommited in 3.3.

Updated converting function uses the calender library to get a local
timezone and then use the GMT datetime to calculate the offset to be
used for calculating, creating, and returning a datetime object.